Update Two Rival Students.cpp

+++ 800/Two Rival Students.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int main()
 {
-	ll t;
+	int t;
 	cin>>t;
 	while(t--)
 	{
